+++ libs/cy-reporter/dev.js
@@ -1,13 +1,5 @@
-const nodemon = require('nodemon')
-const { resolve } = require('path')
-const Runner = require('../../scripts/webpack/dev')
+const Runner = require('../../scripts/runner')
 const compiler = require('./webpack')
 
-const runner = new Runner(compiler, resolve(compiler.outputPath, 'server.js'))
-
-const hasArg = (a) => process.argv.slice(2).includes(a)
-if (hasArg('--build')) {
-  runner.build()
-} else if (hasArg('--watch')) {
-  runner.watch()
-}
+const runner = new Runner(compiler, 'server.js')
+runner.run()
